Improve pure module coverage
Move remaining registration side effects behind explicit, idempotent APIs while preserving legacy import behavior. Correct package side-effect metadata and add bundle and compatibility coverage.
