+++ packages/dev/core/src/Engines/AbstractEngine/abstractEngine.textureLoaders.ts
@@ -1,4 +1,8 @@
-import { _GetCompatibleTextureLoader } from "core/Materials/Textures/Loaders/textureLoaderManager";
-import { AbstractEngine } from "../abstractEngine";
+/**
+ * Re-exports pure implementation and applies runtime side effects.
+ * Import abstractEngine.textureLoaders.pure for tree-shakeable, side-effect-free usage.
+ */
+export * from "./abstractEngine.textureLoaders.pure";
 
-AbstractEngine.GetCompatibleTextureLoader = _GetCompatibleTextureLoader;
+import { RegisterAbstractEngineTextureLoaders } from "./abstractEngine.textureLoaders.pure";
+RegisterAbstractEngineTextureLoaders();
