+++ packages/dev/core/src/Engines/AbstractEngine/pure.ts
@@ -12,4 +12,5 @@ export * from "./abstractEngine.texture.pure";
 export * from "./abstractEngine.alpha.pure";
 export * from "./abstractEngine.views.pure";
 export * from "./abstractEngine.loadFile.pure";
+export * from "./abstractEngine.textureLoaders.pure";
 export * from "./abstractEngine.textureSelector.pure";
