+++ packages/dev/loaders/src/glTF/1.0/pure.ts
@@ -1,3 +1,6 @@
 /** Pure barrel — re-exports only side-effect-free modules */
+export * from "./glTFBinaryExtension.pure";
+export * from "./glTFLoader.pure";
 export * from "./glTFLoaderInterfaces";
 export * from "./glTFLoaderUtils";
+export * from "./glTFMaterialsCommonExtension.pure";
