+++ packages/public/@babylonjs/core/package.json
@@ -39,7 +39,6 @@
         "**/index.js",
         "Shaders/**",
         "ShadersWGSL/**",
-        "**/*SceneComponent.types.js",
         "**/*alphaToCoverage.js",
         "**/*computeShader.js",
         "**/*cubeTexture.js",
@@ -66,22 +65,8 @@
         "**/engine.rawTexture.js",
         "**/inputBlock.js",
         "**/physicsEngineComponent.js",
-        "**/physicsEngineComponent.types.js",
         "**/teleportInBlock.js",
         "**/teleportOutBlock.js",
-        "Animations/*types.js",
-        "Buffers/*types.js",
-        "Cameras/Inputs/*types.js",
-        "Engines/AbstractEngine/*types.js",
-        "Engines/Extensions/*types.js",
-        "Engines/WebGPU/Extensions/*types.js",
-        "Layers/*types.js",
-        "Materials/*types.js",
-        "Materials/Textures/*types.js",
-        "Meshes/*types.js",
-        "Misc/*types.js",
-        "PostProcesses/RenderPipeline/Pipelines/*types.js",
-        "Rendering/*types.js",
         "Actions/action.js",
         "Actions/condition.js",
         "Actions/directActions.js",
@@ -111,7 +96,6 @@
         "Cameras/VR/vrDeviceOrientationGamepadCamera.js",
         "Cameras/arcRotateCamera.js",
         "Cameras/camera.js",
-        "Cameras/camera.types.js",
         "Cameras/deviceOrientationCamera.js",
         "Cameras/flyCamera.js",
         "Cameras/followCamera.js",
@@ -123,21 +107,17 @@
         "Cameras/universalCamera.js",
         "Cameras/virtualJoysticksCamera.js",
         "Collisions/collisionCoordinator.js",
-        "Compute/computeShader.types.js",
         "Culling/Helper/computeShaderBoundingHelper.js",
         "Culling/Helper/transformFeedbackBoundingHelper.js",
         "Culling/Octrees/octreeSceneComponent.js",
         "Culling/ray.js",
         "Debug/debugLayer.js",
-        "Debug/debugLayer.types.js",
         "Debug/physicsViewer.js",
-        "DeviceInput/internalDeviceSourceManager.types.js",
         "Engines/AbstractEngine/abstractEngine.alpha.js",
         "Engines/AbstractEngine/abstractEngine.dom.js",
         "Engines/AbstractEngine/abstractEngine.loadFile.js",
         "Engines/AbstractEngine/abstractEngine.query.js",
         "Engines/AbstractEngine/abstractEngine.renderPass.js",
-        "Engines/AbstractEngine/abstractEngine.scissor.js",
         "Engines/AbstractEngine/abstractEngine.states.js",
         "Engines/AbstractEngine/abstractEngine.stencil.js",
         "Engines/AbstractEngine/abstractEngine.texture.js",
@@ -149,7 +129,6 @@
         "Engines/Extensions/engine.prefilteredCubeTexture.js",
         "Engines/Extensions/engine.transformFeedback.js",
         "Engines/Extensions/engine.uniformBuffer.js",
-        "Engines/Native/Extensions/nativeEngine.cubeTexture.types.js",
         "Engines/Native/validatedNativeDataStream.js",
         "Engines/WebGPU/webgpuClearQuad.js",
         "Engines/WebGPU/webgpuShaderProcessorsWGSL.js",
@@ -157,7 +136,6 @@
         "Engines/nativeEngine.js",
         "Engines/nullEngine.js",
         "Engines/thinEngine.scissor.js",
-        "Engines/thinNativeEngine.types.js",
         "Engines/webgpuEngine.js",
         "FlowGraph/Blocks/Data/Audio/flowGraphGetSoundVolumeBlock.js",
         "FlowGraph/Blocks/Data/Audio/flowGraphIsSoundPlayingBlock.js",
@@ -278,7 +256,6 @@
         "Gamepads/gamepadSceneComponent.js",
         "Gizmos/planeRotationGizmo.js",
         "Helpers/sceneHelpers.js",
-        "Helpers/sceneHelpers.types.js",
         "Layers/effectLayerSceneComponent.js",
         "Layers/glowLayer.js",
         "Layers/highlightLayer.js",
@@ -288,7 +265,6 @@
         "Lights/Clustered/clusteredLightContainer.js",
         "Lights/Clustered/clusteredLightingSceneComponent.js",
         "Lights/Shadows/cascadedShadowGenerator.js",
-        "Lights/areaLight.types.js",
         "Lights/directionalLight.js",
         "Lights/hemisphericLight.js",
         "Lights/lightingVolume.js",
@@ -411,10 +387,8 @@
         "Materials/Node/Blocks/waveBlock.js",
         "Materials/Node/Blocks/worleyNoise3DBlock.js",
         "Materials/Node/nodeMaterial.js",
-        "Materials/Node/nodeMaterial.types.js",
         "Materials/Occlusion/occlusionMaterial.js",
         "Materials/PBR/pbrBaseMaterial.js",
-        "Materials/PBR/pbrMaterial.decalMap.types.js",
         "Materials/PBR/pbrMetallicRoughnessMaterial.js",
         "Materials/PBR/pbrSpecularGlossinessMaterial.js",
         "Materials/Textures/Procedurals/noiseProceduralTexture.js",
@@ -625,10 +599,8 @@
         "Particles/gpuParticleSystem.js",
         "Particles/particleSystem.js",
         "Particles/particleSystemComponent.js",
-        "Particles/particleSystemComponent.types.js",
         "Particles/webgl2ParticleSystem.js",
         "Physics/joinedPhysicsEngineComponent.js",
-        "Physics/joinedPhysicsEngineComponent.types.js",
         "Physics/v1/Plugins/cannonJSPlugin.js",
         "Physics/v1/physicsImpostor.js",
         "PostProcesses/RenderPipeline/Pipelines/defaultRenderingPipeline.js",
@@ -657,7 +629,6 @@
         "PostProcesses/motionBlurPostProcess.js",
         "PostProcesses/passPostProcess.js",
         "PostProcesses/postProcess.js",
-        "PostProcesses/postProcess.types.js",
         "PostProcesses/refractionPostProcess.js",
         "PostProcesses/screenSpaceCurvaturePostProcess.js",
         "PostProcesses/screenSpaceReflectionPostProcess.js",
@@ -668,7 +639,6 @@
         "PostProcesses/volumetricLightScatteringPostProcess.js",
         "PostProcesses/vrMultiviewToSingleviewPostProcess.js",
         "Probes/reflectionProbe.js",
-        "Probes/reflectionProbe.types.js",
         "Rendering/GlobalIllumination/giRSMManager.js",
         "Rendering/IBLShadows/iblShadowsPluginMaterial.js",
         "Rendering/boundingBoxRenderer.js",
@@ -677,7 +647,6 @@
         "Rendering/depthRendererSceneComponent.js",
         "Rendering/edgesRenderer.js",
         "Rendering/fluidRenderer/fluidRenderer.js",
-        "Rendering/fluidRenderer/fluidRenderer.types.js",
         "Rendering/geometryBufferRenderer.js",
         "Rendering/geometryBufferRendererSceneComponent.js",
         "Rendering/iblCdfGeneratorSceneComponent.js",
@@ -706,7 +675,6 @@
         "XR/features/WebXRLightEstimation.js",
         "XR/features/WebXRMeshDetector.js",
         "XR/features/WebXRNearInteraction.js",
-        "XR/features/WebXRNearInteraction.types.js",
         "XR/features/WebXRPlaneDetector.js",
         "XR/features/WebXRRawCameraAccess.js",
         "XR/features/WebXRSpaceWarp.js",
