+++ packages/public/@babylonjs/gui/package.json
