+++ packages/public/@babylonjs/serializers/package.json
